get_data scipt now uses limit and offset parameters to get more than 1000 rows

+++ get_data.py
@@ -31,16 +31,15 @@ headers = {
     'Authorization': f'ApeKey {API_KEY}',  
     'Content-Type': 'application/json'
 }
-
+limit = 1000 # Max limit of data per API Call is 1000
 params = {
     #'mode': 'time'  # Necessary for specific endpoints
+    'limit': limit
 }
 
 # Full url
 url = BASE_URL + results_endpoint
 #print(url)
-# Api request
-response = requests.get(url, headers=headers, params=params)
 
 #print(response)
 
@@ -70,31 +69,46 @@ conn = sqlite3.connect('history.db')
 cursor = conn.cursor()
 
 # Rows counter variable
-rows_counter = 0
+rows_counter = inserted_to_db_rows = 0
+offset = 0
+initial_db_clear = True
 # Get data
 try:
-    data = response.json()  # Parse JSON
-    if data['data']:
-        #print(data['data'])
-        # Delete data from db
-        cursor.execute(sql_query)
-        cursor.execute('DELETE FROM typing_history')
-        # Get headers
-        headers = data['data'][0].keys() 
-        for row in data['data']: 
-            rows_counter+=1
-            # Cast list/dict to str before inserting to db
-            values = [
-            json.dumps(row[header]) if isinstance(row[header], (list,dict)) else row[header]
-            for header in headers
-        ]
-            # Insert into db
-            cursor.execute(f'''
-                INSERT INTO typing_history ({', '.join(headers)})
-                VALUES ({', '.join(['?'] * len(headers))})
-            ''', values)
-    conn.commit()
-    
+    while True:
+        params['offset'] = offset # Add offset to params
+        # Api request
+        response = requests.get(url, headers=headers, params=params)
+        data = response.json()  # Parse JSON
+        if data['data']:
+            #print(data['data'])
+            # Delete data from db
+            if initial_db_clear:
+                cursor.execute(sql_query)
+                cursor.execute('DELETE FROM typing_history')
+                initial_db_clear = False
+                
+            # Get headers
+            headers_row = data['data'][0].keys() 
+            for row in data['data']: 
+                rows_counter+=1
+                # Cast list/dict to str before inserting to db
+                values = [
+                json.dumps(row[header]) if isinstance(row[header], (list,dict)) else row[header]
+                for header in headers_row
+            ]
+                # Insert into db
+                cursor.execute(f'''
+                    INSERT INTO typing_history ({', '.join(headers_row)})
+                    VALUES ({', '.join(['?'] * len(headers_row))})
+                ''', values)
+                inserted_to_db_rows += 1
+            
+            # Commit changes
+            conn.commit()
+            offset += limit # Move forward with data
+        else:
+            break
+        
     # Get db rows count
     db_row_count = 0
     try:
@@ -106,7 +120,7 @@ try:
     
     conn.close()
     with open ('import_status.log', 'a') as file:
-        file.write(f"""{formatDateTime} - Downloaded: {len(data['data'])} rows, Uploaded to db: {rows_counter} rows, db rows count: {db_row_count}\n""")
+        file.write(f"""{formatDateTime} - Downloaded: {rows_counter} rows, Uploaded to db: {inserted_to_db_rows} rows, db rows count: {db_row_count}\n""")
 except Exception as e:
     with open ('logfile.log', 'a') as file:
         file.write(f"""{formatDateTime} Problem with inserting data {e}\n""")
